Factor out in-memory toolpad project
