+++ packages/toolpad-app/pages/api/data/[pageName]/[queryName].ts
@@ -4,10 +4,11 @@ import { ExecFetchResult, SerializedError } from '@mui/toolpad-core';
 import { asArray } from '@mui/toolpad-core/utils/collections';
 import invariant from 'invariant';
 import { withReqResLogs } from '../../../../src/server/logs/withLogs';
-import { execQuery, loadDom } from '../../../../src/server/data';
+import { execQuery } from '../../../../src/server/data';
 import initMiddleware from '../../../../src/server/initMiddleware';
 import * as appDom from '../../../../src/appDom';
 import { errorFrom, serializeError } from '../../../../src/utils/errors';
+import { loadDom } from '../../../../src/server/liveProject';
 
 // Initialize the cors middleware
 const cors = initMiddleware<any>(
