+++ packages/toolpad-app/pages/api/rpc.ts
@@ -1,16 +1,13 @@
 import { NextApiHandler } from 'next';
 import type { IncomingMessage, ServerResponse } from 'http';
 import superjson from 'superjson';
-import { execQuery, dataSourceFetchPrivate, loadDom, saveDom } from '../../src/server/data';
+import { execQuery, dataSourceFetchPrivate } from '../../src/server/data';
 import { getLatestToolpadRelease } from '../../src/server/getLatestRelease';
 import { hasOwnProperty } from '../../src/utils/collections';
 import { errorFrom, serializeError } from '../../src/utils/errors';
 import logger from '../../src/server/logs/logger';
-import {
-  createComponent,
-  getDomFingerprint,
-  openCodeComponentEditor,
-} from '../../src/server/localMode';
+import { createComponent, openCodeComponentEditor } from '../../src/server/localMode';
+import { getDomFingerprint, loadDom, saveDom } from '../../src/server/liveProject';
 
 export interface Method<P extends any[] = any[], R = any> {
   (...params: P): Promise<R>;
