+++ packages/toolpad-app/src/server/liveProject.ts
@@ -1,0 +1,30 @@
+import chalk from 'chalk';
+import * as appDom from '../appDom';
+import { initProject } from './localMode';
+
+// eslint-disable-next-line no-underscore-dangle
+(globalThis as any).__project__ ??= initProject().catch((err) => {
+  console.error(`${chalk.red('error')} - Failed to intialize Toolpad`);
+  console.error(err);
+  process.exit(1);
+});
+
+export async function waitForInit(): Promise<ReturnType<typeof initProject>> {
+  // eslint-disable-next-line no-underscore-dangle
+  return (globalThis as any).__project__;
+}
+
+export async function getDomFingerprint() {
+  const project = await waitForInit();
+  return project.getDomFingerPrint();
+}
+
+export async function saveDom(newDom: appDom.AppDom): Promise<{ fingerprint: number }> {
+  const project = await waitForInit();
+  return project.saveDom(newDom);
+}
+
+export async function loadDom(): Promise<appDom.AppDom> {
+  const project = await waitForInit();
+  return project.loadDom();
+}
